+++ oereb_client/static/src/util/error.js
@@ -4,3 +4,10 @@ export class NoDataError extends Error {
     this.name = "NoDataError";
   }
 }
+
+export class TooManyRequestsError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "TooManyRequestsError";
+  }
+}